+++ src/vfvm_testfunctions.jl
@@ -8,17 +8,17 @@ test functions for boundary flux calculations.
 
 $(TYPEDFIELDS)
 """
-mutable struct TestFunctionFactory{Tv}
+mutable struct TestFunctionFactory
 
     """
     Original system
     """
-    system::AbstractSystem{Tv}
+    system::AbstractSystem
 
     """
     Test function system
     """
-    tfsystem::DenseSystem{Tv}
+    tfsystem::DenseSystem
 
 
     """
@@ -35,7 +35,7 @@ $(TYPEDSIGNATURES)
 
 Constructor for TestFunctionFactory from System
 """
-function TestFunctionFactory(system::AbstractSystem{Tv}; control=SolverControl()) where Tv
+function TestFunctionFactory(system::AbstractSystem; control=SolverControl())
     physics=Physics( 
         flux=function(f,u,edge)
         f[1]=u[1]-u[2]
@@ -46,7 +46,7 @@ function TestFunctionFactory(system::AbstractSystem{Tv}; control=SolverControl()
     )
     tfsystem=System(system.grid,physics,unknown_storage=:dense)
     enable_species!(tfsystem,1,[i for i=1:num_cellregions(system.grid)])
-    return TestFunctionFactory{Tv}(system,tfsystem,control)
+    return TestFunctionFactory(system,tfsystem,control)
 end
 
 
@@ -59,7 +59,7 @@ Create testfunction which has Dirichlet zero boundary conditions  for boundary
 regions in bc0 and Dirichlet one boundary conditions  for boundary
 regions in bc1.
 """
-function testfunction(factory::TestFunctionFactory{Tv}, bc0, bc1) where Tv
+function testfunction(factory::TestFunctionFactory, bc0, bc1)
 
     u=unknowns(factory.tfsystem)
     f=unknowns(factory.tfsystem)
@@ -100,7 +100,7 @@ $(SIGNATURES)
 
 Calculate test function integral for transient solution.
 """
-function integrate(system::AbstractSystem{Tv},tf::Vector{Tv},U::AbstractMatrix{Tv},
+function integrate(system::AbstractSystem,tf,U::AbstractMatrix{Tv},
                    Uold::AbstractMatrix{Tv}, tstep::Real) where {Tv}
     grid=system.grid
     nspecies=num_species(system)
@@ -111,6 +111,7 @@ function integrate(system::AbstractSystem{Tv},tf::Vector{Tv},U::AbstractMatrix{T
     storold=zeros(Tv,nspecies)
     tstepinv=1.0/tstep
 
+    #!!! params etc 
     physics=system.physics
     node=Node(system)
     bnode=BNode(system)
@@ -172,7 +173,7 @@ $(SIGNATURES)
 
 Calculate test function integral for steady state solution.
 """
-integrate(system::AbstractSystem,tf::Vector{Tv},U::AbstractMatrix{Tv}) where Tv =integrate(system,tf,U,U,Inf)
+integrate(system::AbstractSystem,tf::Vector{Tv},U::AbstractMatrix{Tu}) where {Tu,Tv} =integrate(system,tf,U,U,Inf)
 
 
 
@@ -182,7 +183,7 @@ $(SIGNATURES)
 
 Steady state part of test function integral.
 """
-function integrate_stdy(system::AbstractSystem,tf::Vector{Tv},U::AbstractArray{Tu,2}) where {Tu,Tv,Ti}
+function integrate_stdy(system::AbstractSystem,tf::Vector{Tv},U::AbstractArray{Tu,2}) where {Tu,Tv}
     grid=system.grid
     nspecies=num_species(system)
     integral=zeros(Tu,nspecies)
@@ -241,7 +242,7 @@ $(SIGNATURES)
 
 Calculate transient part of test function integral.
 """
-function integrate_tran(system::AbstractSystem,tf::Vector{Tv},U::AbstractArray{Tu,2}) where {Tu,Tv,Ti}
+function integrate_tran(system::AbstractSystem,tf::Vector{Tv},U::AbstractArray{Tu,2}) where {Tu,Tv}
     grid=system.grid
     nspecies=num_species(system)
     integral=zeros(Tu,nspecies)